backend/requirements.txt
openai
pdfplumber
python-dotenv
fastapi
uvicorn[standard]
python-multipart
